Refactor setup.py into utils.py to simplify scripts, modify mnist calibration data class to take in the name of the first input to the model
